Move focal-state into separate packages
