+++ package.json
@@ -2,10 +2,10 @@
   "private": true,
   "name": "focal",
   "scripts": {
-    "build": "yarn workspace @grammarly/focal build",
+    "build": "yarn workspace @grammarly/focal-utils build && yarn workspace @grammarly/focal-state build && yarn workspace @grammarly/focal build",
     "package": "yarn workspace @grammarly/focal pack",
     "release": "yarn workspace @grammarly/focal release",
-    "test": "yarn workspace @grammarly/focal test && yarn workspace focal-todomvc build && yarn workspace focal-examples build && yarn workspace focal-manual-tests build",
+    "test": "yarn workspace @grammarly/focal-state test && yarn workspace focal-todomvc build && yarn workspace focal-examples build && yarn workspace focal-manual-tests build",
     "postinstall": "yarn build"
   },
   "devDependencies": {
@@ -14,6 +14,8 @@
     "typescript": "3.3.3"
   },
   "workspaces": [
+    "packages/focal-state",
+    "packages/focal-utils",
     "packages/focal",
     "packages/test",
     "packages/examples/*"
