+++ packages/focal-state/src/atom/base.ts
