+++ packages/focal-state/tsconfig.json
@@ -1,0 +1,25 @@
+{
+  "compilerOptions": {
+    "declaration": true,
+    "target": "es5",
+    "sourceMap": true,
+    "emitDecoratorMetadata": true,
+    "experimentalDecorators": true,
+    "rootDir": ".",
+    "noEmitOnError": true,
+    "noImplicitReturns": true,
+    "noImplicitAny": true,
+    "noFallthroughCasesInSwitch": true,
+    "noUnusedLocals": false,
+    "noUnusedParameters": false,
+    "moduleResolution": "node",
+    "jsx": "react",
+    "strictNullChecks": true,
+    "noImplicitThis": true,
+    "lib": ["es6", "es2015", "dom", "scripthost"]
+  },
+  "include": [
+    "src/**/*.ts*",
+    "test/**/*.ts*"
+  ]
+}
