+++ packages/focal-utils/src/equals.ts
