+++ packages/focal-utils/src/index.ts
@@ -1,0 +1,1 @@
+export { equals as structEq } from './equals'
